requirements.txt

python-dotenv
joblib
kaggle
numpy
pandas
scikit-learn
tensorflow
flask